src/requirements.txt
matplotlib==3.2.2
numpy==1.18.5
opencv_python==4.5.1.48
pandas==1.0.5
Pillow==9.0.1
rglob==1.7
scikit_learn==1.0.2
Shapely==1.7.1
tensorflow==2.2.0
tqdm==4.47.0
